created fragment MyAlertDialogFragment.java, added MyAlertDialogFragment to button Exit in OptionsMenu
